pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arbitrage-bot-dashboard"
version = "1.0.0"
description = "Real-time dashboard for monitoring and controlling the arbitrage bot"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {file = "LICENSE"}
requires-python = ">=3.9"
dependencies = [
    "flask==2.3.2",
    "flask-socketio==5.3.6",
    "flask-cors==4.0.0",
    "eventlet==0.33.3",
    "python-dotenv==1.0.0",
    "web3==6.5.0",
    "eth-account==0.9.0",
    "requests==2.31.0",
    "tenacity==8.2.2",
    "pandas",
    "numpy"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0"
]

[tool.setuptools]
packages = ["dashboard"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=dashboard"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["dashboard"]
omit = [
    "tests/*",
    "**/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[project.urls]
Homepage = "https://github.com/yourusername/arbitrage-bot"
Documentation = "https://github.com/yourusername/arbitrage-bot/docs"
Repository = "https://github.com/yourusername/arbitrage-bot.git"
Issues = "https://github.com/yourusername/arbitrage-bot/issues"

[project.scripts]
arbitrage-dashboard = "dashboard.app:main"
